Started to add tsne pipeline. Gpu code missing.

+++ PCViewer/test/ranges_test.cpp
@@ -42,17 +42,17 @@ int test_find_if(){
 
 int test_rev_size_range(){
     std::vector<int> v(16);
-    for(int i: util::rev_size_range(v))
+    for(size_t i: util::rev_size_range(v))
         if(i < 0)
             return test_result::error;
     structures::change_tracker<std::vector<int>> c;
     c() = {16, 10, 13};
-    for(int i: util::rev_size_range(c.read()))
+    for(size_t i: util::rev_size_range(c.read()))
         if(i < 0)
             return test_result::error;
 
     std::map<int, std::vector<int>> m{{2, {1,8,7}}};
-    for(int i: util::rev_size_range(m[2]))
+    for(size_t i: util::rev_size_range(m[2]))
         if(i < 0)
             return test_result::error;
     return test_result::success;
